Version Info
Get version and other info at build time.
