+++ .gitignore
@@ -2,3 +2,5 @@
 node_modules/
 test/output/
 dist/
+
+.DS_Store
