+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "grumptech-fs-hasher",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "description": "Generates hash digests of file system folders and their contents.",
   "main": "dist/grumptech-fs-hasher.js",
   "type": "module",
@@ -63,11 +63,13 @@
     "eslint-plugin-promise": "^6.0.1",
     "jest": "^29.1.2",
     "jsdoc": "^3.6.11",
+    "string-replace-loader": "^3.1.0",
     "webpack": "^5.74.0",
-    "webpack-cli": "^4.10.0",
+    "webpack-cli": "^5.0.1",
     "webpack-node-externals": "^3.0.0"
   },
   "dependencies": {
+    "copy-webpack-plugin": "^11.0.0",
     "debug": "^4.3.4",
     "is-it-check": "^1.0.11",
     "run-script-os": "^1.1.5"
